\name{rlips.copy}
\alias{rlips.copy}

\title{Copy RLIPS environment}

\description{
Copy RLIPS environment into a new environment}

\usage{
rlips.copy(oenv)
}

\arguments{
\item{oenv}{RLIPS environment}
}

\value{RLIPS environment which is a copy of \code{oenv}.}

\author{Mikko Orispaa (University of Oulu, Finland) \email{mikko.orispaa@oulu.fi}}

%\examples{}

\keyword{algebra}
