\name{rlips.dispose}
\alias{rlips.dispose}

\title{Delete rlips environment}

\description{Deletes the rlips enviroment's internal variables and the files
used by the environment.}

\usage{rlips.dispose(e)}

\arguments{
	\item{e}{rlips environment to be deleted}
}

\value{None. \code{e} will be an empty environment afterwards.}

\details{\code{rlips.dispose} will delete all variables in rlips environment \code{h}.}

\author{Mikko Orispaa (University of Oulu, Finland) \email{mikko.orispaa@oulu.fi}}

\keyword{algebra}
